Moving GenericNetwork to network protocol
