+++ mikeio1d/experimental/_network_mapper.py
@@ -13,7 +13,12 @@ from typing import Any, Callable
 
 from mikeio1d import Res1D
 from mikeio1d.result_network import ResultNode, ResultReach, ResultGridPoint
-from mikeio1d.experimental._network_protocol import NetworkEdge, NetworkNode, EdgeBreakPoint
+from mikeio1d.experimental._network_protocol import (
+    NetworkEdge,
+    NetworkNode,
+    EdgeBreakPoint,
+    GenericNetwork,
+)
 
 
 class NetworkBackend(Enum):
@@ -156,81 +161,6 @@ def parse_res1d_network(res: Any) -> dict[str, NetworkEdge]:
 
     network = read_res1d_network(res)
     return {reach_id: parse_reach(reach) for reach_id, reach in network.reaches.items()}
-
-
-class GenericNetwork:
-    """Generic network structure."""
-
-    def __init__(self, graph: nx.Graph):
-        self._graph = graph.copy()
-        self._df = self._build_dataframe()
-
-    def _build_dataframe(self) -> pd.DataFrame:
-        df = pd.concat({k: v["data"] for k, v in self._graph.nodes.items()}, axis=1)
-        df.columns = df.columns.set_names(["node", "quantity"])
-        df.index.name = "time"
-        return df.copy()
-
-    def to_dataframe(self, sel: str | None = None) -> pd.DataFrame:
-        """Dataframe using node ids as column names.
-
-        It will be multiindex unless 'sel' is passed.
-
-        Parameters
-        ----------
-        sel : Optional[str], optional
-            Quantity to select, by default None
-
-        Returns
-        -------
-        pd.DataFrame
-            Timeseries contained in graph nodes
-        """
-        df = self._df.copy()
-        if sel is None:
-            return df
-        else:
-            df.attrs["quantity"] = sel
-            return df.reorder_levels(["quantity", "node"], axis=1).loc[:, sel]
-
-    def to_dataset(self) -> xr.Dataset:
-        """Dataset using node ids as coords.
-
-        Returns
-        -------
-        xr.Dataset
-            Timeseries contained in graph nodes
-        """
-        df = self.to_dataframe()
-        df = df.reorder_levels(["quantity", "node"], axis=1).melt(ignore_index=False)
-
-        duplicate_check = df.reset_index().duplicated()
-        if duplicate_check.any():
-            raise ValueError("Duplicated values found")
-
-        df = df.pivot_table(
-            index=["time", "node"],
-            columns="quantity",
-            values="value",
-            aggfunc="first",
-        )
-        return df.to_xarray()
-
-    @property
-    def graph(self) -> nx.Graph:
-        """Graph of the network."""
-        return self._graph
-
-    @property
-    def quantities(self) -> list[str]:
-        """Quantities present in data.
-
-        Returns
-        -------
-        List[str]
-            List of quantities
-        """
-        return list(self.to_dataframe().columns.get_level_values(1).unique())
 
 
 class NetworkMapper:
